requirements/runtime.txt
openmm>=7.7
pdbfixer
mpi4py
openmmforcefields
torch
torch-cluster
torch-scatter
torch-sparse
torch-spline-conv
torch-geometric
pymol
yapf==0.32.0
addict==2.4.0
prettytable==3.6.0
easydict==1.10
omegaconf
tqdm
typing-extensions
networkx
opencv-python
matplotlib
seaborn
requests
pandas
numba
scikit-learn>=1.1.3
scipy>=1.12.0
e3nn
prody<=2.4.0
biopython==1.80
rdkit>=2023.03.2
MDAnalysis
einops
py3Dmol
nglview
joblib==1.1.0
pandarallel==1.6.3
spyrmsd==0.5.2
tensorboardX
tensorboard
torchmetrics
torchtyping
typeguard
gpustat
dm-tree
lmdb==1.4.1
wandb==0.13.3
filelock
posebusters
prefetch_generator
mdtraj
docutils